chore: Setup new test suites - Update planx-core - Move shared helpers - Scope test lifecycle hooks - Add JWT helper methods
