+++ e2e/tests/api-driven/src/client.ts
@@ -1,5 +1,7 @@
 import assert from "node:assert";
 import { CoreDomainClient } from "@opensystemslab/planx-core";
+import { TEST_EMAIL } from "../../ui-driven/src/helpers";
+import { buildJWT } from "./jwt";
 
 // check env variables are defined
 assert(process.env.HASURA_GRAPHQL_ADMIN_SECRET);
@@ -11,6 +13,21 @@ const targetURL = process.env.HASURA_GRAPHQL_URL!.replace(
 );
 
 export const $admin = new CoreDomainClient({
-  hasuraSecret: process.env.HASURA_GRAPHQL_ADMIN_SECRET!,
+  auth: { adminSecret: process.env.HASURA_GRAPHQL_ADMIN_SECRET! },
   targetURL,
 });
+
+/**
+ * Get client authorised to the permissions level of the test user
+ */
+export const getClient = async () => {
+  const jwt = await buildJWT(TEST_EMAIL);
+  if (!jwt) throw Error("Unable to generate JWT for test user");
+
+  const client = new CoreDomainClient({
+    targetURL: process.env.HASURA_GRAPHQL_URL!,
+    auth: { jwt },
+  });
+
+  return client;
+};
